Add Perl's IPC::Run as a build-time dependency
IPC::Run is a run-time requirement for chronic.  To allow the
requirement to be detected properly by Cygport, make sure it's installed
at compile time.

This is preferable to just listing it as a package dependency as it
means the requirements will automatically update if the Cygwin package
name changes or the dependency gets removed.